simulation/requirements.txt
pymongo==4.7.2
python-dotenv==1.0.1
faker==25.2.0
schedule==1.2.2